no need for separate exceptions module
