--- nameko/cli/exceptions.py
@@ -1,2 +1,0 @@
-class CommandError(Exception):
-    """Raise from subcommands to report error back to the user"""
